* Updated kdevelop-related files * Removed unecessary members at actions and commands classes * Replaced getAction looping through several maps with single boolean at internalUseItem * Beds may now have external scripts before player falls asleep- possibility to deny sleep due to scripted "conditions" * Added RET_YOUARENOTTHEOWNER, as 'You are not the owner.' message wasn't appearing * Fixed compiler warnings Hopefuly didn't cause any compiler error, because couldn't compile...

+++ otserv/trunk/fileloader.h
@@ -226,7 +226,7 @@ public:
 		if(!GET_USHORT(str_len)){
 			return false;
 		}
-		if(size() < str_len){
+		if(size() < (int32_t)str_len){
 			return false;
 		}
 		str = new char[str_len+1];
@@ -245,7 +245,7 @@ public:
 		if(!GET_ULONG(str_len)){
 			return false;
 		}
-		if(size() < str_len){
+		if(size() < (int32_t)str_len){
 			return false;
 		}
 		str = new char[str_len+1];
@@ -260,7 +260,7 @@ public:
 	inline bool GET_NSTRING(unsigned short str_len, std::string& ret){
 		char* str;
 
-		if(size() < str_len){
+		if(size() < (int32_t)str_len){
 			return false;
 		}
 		str = new char[str_len+1];
